OSIS-1246 Rename file bibliography to teaching_material
